Make sure that the role of main.rb is only to start the app and ask the user for the initial choice of options
